LMM_project/memory_store/index_manager.py
from typing import Dict, List, Optional, Union, Any, Tuple
import numpy as np
from pathlib import Path
import os
import faiss
import torch
import json
import time
from enum import Enum

from pydantic import BaseModel, Field, field_validator
from models.memory_models import FAISSIndexConfig, EmbeddingVector

# TODO: Define IndexType enum with supported FAISS index types
# TODO: Create MetricType enum (L2, INNER_PRODUCT, etc.)
# TODO: Implement IndexState model for serialization

# TODO: Implement FAISSIndexManager class:
#   - __init__ with config and GPU detection
#   - create_index method for new indices
#   - save_index method for persistence (Windows-compatible)
#   - load_index method for loading from disk
#   - add_vectors method to add new embeddings
#   - search method for similarity searches
#   - update_vectors method for modifying existing entries
#   - remove_vectors method for deletion
#   - get_index_stats method for monitoring
#   - optimize_for_inference method for inference preparation
#   - move_to_gpu and move_to_cpu methods for resource management

# TODO: Create CUDAHelper for Windows CUDA optimization:
#   - detect_cuda_availability method
#   - get_optimal_gpu_id method
#   - manage_gpu_resources method
#   - release_gpu_resources method

# TODO: Implement BatchProcessor for efficient vector operations:
#   - process_batch method for chunked operations
#   - optimize_batch_size method based on available memory

# TODO: Create IndexRegistry for managing multiple indices:
#   - register_index method
#   - get_index method
#   - list_indices method
#   - remove_index method

# TODO: Add Windows-specific error handling for CUDA issues
# TODO: Implement optimization for Windows memory constraints
# TODO: Create resource monitoring and cleanup mechanisms